+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 1. [](#new)
     * Added directory configuration option for custom admin templates
     * Added `Flex Accounts (Admin)` type to administer user accounts in Flex independently from Grav system setting
+    * Added `Flex Pages (Admin)` type to administer pages in Flex independently from Grav system setting
 1. [](#bugfix)
     * Flex Pages: Fixed default language not being translated in both `translatedLanguages()` and `untranslatedLanguages()` results
     * Flex Pages: Language interface compatibility fixes