+++ CHANGELOG.md
@@ -12,6 +12,7 @@
     * Flex Pages: Language interface compatibility fixes
     * Flex Pages: Fixed frontend issues with plugin events [#5](https://github.com/trilbymedia/grav-plugin-flex-objects/issues/5)
     * Flex Pages: Fixed `filePathClean()` and `filePathClean()` not returning file for folder
+    * Flex Pages: Fixed multiple multi-language related issues in admin [#10](https://github.com/trilbymedia/grav-plugin-flex-objects/issues/10)
 
 # v1.0.0-beta.10
 ## 10/03/2019